+++ resources/share.rb
@@ -16,27 +16,90 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-property :share_name, String, name_property: true
-property :comment, String
-property :path, String, required: false
-property :valid_users, String, default: ''
-property :force_group, String, default: ''
-property :browseable, String, default: 'yes', equal_to: %w(yes no)
-property :guest_ok, String, default: 'no', equal_to: %w(yes no)
-property :printable, String, default: 'no', equal_to: %w(yes no)
-property :write_list, Array, required: false
-property :create_mask, String, default: '0744', required: false
-property :directory_mask, String, default: '0755', required: false
-property :read_only, String, default: 'no', equal_to: %w(yes no)
-property :create_directory, [true, false], default: true
-property :options, Hash, default: {}
-property :config_file, String, default: lazy {
-  if node['platform_family'] == 'smartos' # rubocop: disable ChefStyle/UsePlatformHelpers
-    '/opt/local/etc/samba/smb.conf'
-  else
-    '/etc/samba/smb.conf'
-  end
-}
+
+unified_mode true
+
+property :share_name,
+        String,
+        name_property: true,
+        description: 'The name of the share'
+
+property :comment,
+        String,
+        description: 'Comment string to associate with the new share'
+
+property :path,
+        String,
+        required: false,
+        description: 'Path to directory to share'
+
+property :valid_users,
+        String,
+        description: 'A string of allowed users'
+
+property :force_group,
+        String,
+        description: 'Force ownership of files on the share to specified group'
+
+property :force_user,
+        String,
+        description: 'Force ownership of files on the share to specified user'
+
+property :browseable,
+        [true, false, String],
+        default: true,
+        coerce: proc { |p| p ? 'yes' : 'no' },
+        description: 'Controls whether this share is seen in the list of available shares in a net view and in the browse list'
+
+property :guest_ok,
+        [true, false, String],
+        default: false,
+        coerce: proc { |p| p ? 'yes' : 'no' },
+        description: 'Allow anoymous access to the share'
+
+property :printable,
+        [true, false, String],
+        default: false,
+        coerce: proc { |p| p ? 'yes' : 'no' },
+        description: 'If set to yes, then clients may open, write to and submit spool files on the directory specified for the service'
+
+property :write_list,
+        Array,
+        required: false,
+        description: 'An array of Unix users allowed to write to the share'
+
+property :create_mask,
+        String,
+        default: '0744',
+        required: false,
+        description: 'Create mask for directory'
+
+property :directory_mask,
+        String,
+        default: '0755',
+        required: false,
+        description: 'Mask for directory'
+
+property :read_only,
+        [true, false, String],
+        default: false,
+        coerce: proc { |p| p ? 'yes' : 'no' },
+        description: 'Whether files on the share are writeable'
+
+property :create_directory,
+        [true, false, String],
+        default: true,
+        description: 'Whether to create the new share directory on disk'
+
+property :options,
+        Hash,
+        default: {},
+        description: 'A hash of extra options to pass to the configuration file'
+
+property :config_file,
+        String,
+        default: lazy { platform?('smartos') ? '/opt/local/etc/samba/smb.conf' : '/etc/samba/smb.conf' },
+        description: 'Path to the samba configuration file'
 
 action :add do
   # We need to force both the server template and the
@@ -55,6 +118,7 @@ action :add do
       variables[:shares][new_resource.share_name]['directory mask'] = new_resource.directory_mask
       variables[:shares][new_resource.share_name]['read only'] = new_resource.read_only
       variables[:shares][new_resource.share_name]['valid users'] = new_resource.valid_users
+      variables[:shares][new_resource.share_name]['force user'] = new_resource.force_user
       variables[:shares][new_resource.share_name]['force group'] = new_resource.force_group
       variables[:shares][new_resource.share_name]['browseable'] = new_resource.browseable
       new_resource.options.each do |key, value|