+++ resources/user.rb
@@ -16,11 +16,31 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-property :password, String
-property :comment, String
-property :home, String, default: lazy { ::File.join('/home/', name) }
-property :shell, String, default: '/bin/bash'
-property :manage_home, [true, false], default: true
+unified_mode true
+
+property :password,
+        String,
+        sensitive: true,
+        description: 'User password for samba and the system'
+
+property :comment,
+        String,
+        description: 'One (or more) comments about the user'
+
+property :home,
+        String,
+        default: lazy { ::File.join('/home/', name) },
+        description: 'Users home'
+
+property :shell,
+        String,
+        default: '/bin/bash',
+        description: 'User shell to set, e.g. /bin/sh, /sbin/nologin'
+
+property :manage_home,
+        [true, false],
+        default: true,
+        description: 'Whether to manage the users home directory location'
 
 def load_current_value
   @smbuser = Chef::Resource::SambaUser.new(new_resource.name)
@@ -36,7 +56,7 @@ end
 
 action :create do
   user new_resource.name do
-    password generate_system_password
+    password new_resource.password
     comment new_resource.comment
     home new_resource.home
     shell new_resource.shell
@@ -76,14 +96,5 @@ action :delete do
     execute "Delete #{new_resource.name}" do
       command "smbpasswd -x #{new_resource.name}"
     end
-  end
-end
-
-action_class.class_eval do
-  def generate_system_password
-    system_password = \
-      new_resource.password.crypt('$6$' + SecureRandom.random_number(36**8).to_s(36))
-    Chef::Log.debug "SC: generated system password: #{system_password}"
-    system_password
   end
 end