+++ test/fixtures/cookbooks/test/recipes/server.rb
@@ -1,7 +1,6 @@
 apt_update 'update'
 
-samba_server 'server' do
-  server_string 'fat'
+samba_server 'foxtrot' do
   workgroup 'HOME'
   interfaces 'lo 127.0.0.1 enp* eth*'
   hosts_allow '0.0.0.0/0'
@@ -28,8 +27,8 @@ end
 samba_share 'first_share' do
   comment 'exported share 1'
   path '/srv/export'
-  guest_ok 'no'
-  printable 'no'
+  guest_ok true
+  printable false
   write_list ['test_user_1']
   create_mask '0644'
   directory_mask '0775'
@@ -39,8 +38,8 @@ end
 samba_share 'second_share' do
   comment 'exported share 2'
   path '/srv/export_2'
-  guest_ok 'no'
-  printable 'no'
+  guest_ok false
+  printable false
   write_list ['test_user_2']
   create_mask '0644'
   directory_mask '0775'